requirements.txt
streamlit==1.32.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-core>=0.1.16
openai==1.12.0
python-dotenv==1.0.0
pydantic>=2.5.2
